+++ CircleAttack.cpp
@@ -3,7 +3,7 @@
 
 #include <random>
 
-CircleAttack::CircleAttack(Surface* screen, int duration, int amountOfProj, int amountOfCircles)
+CircleAttack::CircleAttack(Surface* screen, float duration, int amountOfProj, int amountOfCircles)
     : BaseAttack(screen, duration, 0), projectilesAmount(amountOfProj), circleAmount(amountOfCircles), radius(200.f)
 {
 	if (circleAmount == 0)
